requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
black==23.12.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
environs==10.0.0
exceptiongroup==1.2.0
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pluggy==1.3.0
pyaes==1.6.1
pyasn1==0.5.1
pytest==7.4.3
python-dotenv==1.0.0
python-pytun==2.4.1
python-pytuntap==1.0.5
requests==2.31.0
rsa==4.9
Telethon==1.33.1
tomli==2.0.1
typing-extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
